Tucker product.

Performs the Tucker product between an array and a list of matrices.

atrans(A, B)

Arguments

A
An array of dimension \(K\).

B
A list of matrices of length \(K\). It must be that ncol(B[[k]]) == dim(A)[k].

Details

The Tucker product between a list of matrices B and an array A is formally equivalent to performing the \(k\)-mode product between A and each list element in B. For example, if the dimension of A is three, then atrans(A,B) = amprod(amprod(amprod(A,B[[1]],1),B[[2]],2),B[[3]],3). The ordering of this \(k\)-mode product does not matter. See Kolda and Bader (2009) for details.

References

Kolda, T. G., & Bader, B. W. (2009). Tensor decompositions and applications. SIAM review, 51(3), 455-500.

See also

Examples

A <- array(1:8, dim = c(2,2,2))
B <- list()
B[[1]] <-matrix(1:4, nrow = 2)
B[[2]] <- matrix(1:6, nrow = 3)
B[[3]] <- matrix(1:2, nrow = 1)
atrans(A,B)

#> , , 1
#> 
#>      [,1] [,2] [,3]
#> [1,]  361  491  621
#> [2,]  534  726  918
#>