Calculate the incredible SVD (ISVD).

The ISVD is a generalization of the SVD to tensors. It is derived from the incredible HOLQ.

get_isvd(x_holq)

Arguments

x_holq
The output from holq.

Value

l A numeric.

U A list of orthogonal matrices.

D A list of diagonal matrices with positive diagonal entries and unit determinant. The diagonal entries are in descending order.

V A scaled all-orthonormal array.

Details

Let sig * atrans(Z, L) be the HOLQ of X. Then the ISVD calculates the SVD of each L[[i]], call it U[[i]] %*% D[[i]] %*% t(W[[i]]). It then returns l = sig, U, D, and V = atrans(Z, W). These values have the property that X is equal to l * atrans(atrans(V, D), U), up to numerical precision. V is also scaled all-orthonormal.

For more details on the ISVD, see Gerard and Hoff (2016).

References

Gerard, D., & Hoff, P. (2016). A higher-order LQ decomposition for separable covariance models. Linear Algebra and its Applications, 505, 57-84. https://doi.org/10.1016/j.laa.2016.04.033 http://arxiv.org/pdf/1410.1094v1.pdf

Examples

#Generate random data.
p <- c(4,4,4)
X <- array(stats::rnorm(prod(p)), dim = p)

#Calculate HOLQ, then ISVD
holq_x <- holq(X)

#> Scale Diff =  0.1970162 
#>      Scale =  6.660327 
#> 
#> Scale Diff =  0.0490614 
#>      Scale =  6.348844 
#> 
#> Scale Diff =  0.01107 
#>      Scale =  6.279331 
#> 
#> Scale Diff =  0.004510989 
#>      Scale =  6.251133 
#> 
#> Scale Diff =  0.001876543 
#>      Scale =  6.239424 
#> 
#> Scale Diff =  0.0007692939 
#>      Scale =  6.234628 
#> 
#> Scale Diff =  0.0003119185 
#>      Scale =  6.232684 
#> 
#> Scale Diff =  0.0001263392 
#>      Scale =  6.231896 
#> 
#> Scale Diff =  5.149719e-05 
#>      Scale =  6.231575 
#> 
#> Scale Diff =  2.116027e-05 
#>      Scale =  6.231444 
#> 
#> Scale Diff =  8.747985e-06 
#>      Scale =  6.231389 
#> 
#> Scale Diff =  3.628263e-06 
#>      Scale =  6.231366 
#> 
#> Scale Diff =  1.506393e-06 
#>      Scale =  6.231357 
#> 
#> Scale Diff =  6.254167e-07 
#>      Scale =  6.231353 
#> 
#> Scale Diff =  2.596463e-07 
#>      Scale =  6.231352 
#> 
#> Scale Diff =  1.07864e-07 
#>      Scale =  6.231351 
#> 
#> Scale Diff =  4.488738e-08 
#>      Scale =  6.231351 
#> 
#> Scale Diff =  1.873511e-08 
#>      Scale =  6.23135 
#> 
#> Scale Diff =  7.852003e-09 
#>      Scale =  6.23135 
#> 
#> Scale Diff =  3.307732e-09 
#>      Scale =  6.23135 
#> 
#> Scale Diff =  1.401634e-09 
#>      Scale =  6.23135 
#> 
#> Scale Diff =  5.977341e-10 
#>      Scale =  6.23135 
#> 

isvd_x <- get_isvd(holq_x)
l <- isvd_x$l
U <- isvd_x$U
D <- isvd_x$D
V <- isvd_x$V

#Recover X
trim(X - l * atrans(atrans(V, D), U))

#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    0    0    0    0
#> [2,]    0    0    0    0
#> [3,]    0    0    0    0
#> [4,]    0    0    0    0
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    0    0    0    0
#> [2,]    0    0    0    0
#> [3,]    0    0    0    0
#> [4,]    0    0    0    0
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    0    0    0    0
#> [2,]    0    0    0    0
#> [3,]    0    0    0    0
#> [4,]    0    0    0    0
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    0    0    0    0
#> [2,]    0    0    0    0
#> [3,]    0    0    0    0
#> [4,]    0    0    0    0
#> 

#V is scaled all-orthonormal
trim(mat(V, 1) %*% t(mat(V, 1)), epsilon = 10^-5)

#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 0.2500029 0.0000000 0.0000000 0.0000000
#> [2,] 0.0000000 0.2500022 0.0000000 0.0000000
#> [3,] 0.0000000 0.0000000 0.2500011 0.0000000
#> [4,] 0.0000000 0.0000000 0.0000000 0.2499938

trim(mat(V, 2) %*% t(mat(V, 2)), epsilon = 10^-5)

#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 0.2500008 0.0000000 0.0000000 0.0000000
#> [2,] 0.0000000 0.2500012 0.0000000 0.0000000
#> [3,] 0.0000000 0.0000000 0.2499979 0.0000000
#> [4,] 0.0000000 0.0000000 0.0000000 0.2500001

trim(mat(V, 3) %*% t(mat(V, 3)), epsilon = 10^-5)

#>      [,1] [,2] [,3] [,4]
#> [1,] 0.25 0.00 0.00 0.00
#> [2,] 0.00 0.25 0.00 0.00
#> [3,] 0.00 0.00 0.25 0.00
#> [4,] 0.00 0.00 0.00 0.25