Calculate the (truncated) higher-order SVD (HOSVD).

Calculates the left singular vectors of each matrix unfolding of an array, then calculates the core array. The resulting output is a Tucker decomposition.

hosvd(Y, r = NULL)

Arguments

Y

An array of numerics.

r

A vector of integers. The rank of the truncated HOSVD.

Value

U A list of matrices with orthonormal columns. Each matrix contains the mode-specific singular vectors of its mode.

S An all-orthogonal array. This is the core array from the HOSVD.

Details

If r is equal to the rank of Y, then Y is equal to atrans(S, U), up to numerical accuracy.

More details on the HOSVD can be found in De Lathauwer et. al. (2000).

References

De Lathauwer, L., De Moor, B., & Vandewalle, J. (2000). A multilinear singular value decomposition. SIAM journal on Matrix Analysis and Applications, 21(4), 1253-1278.

Author

Peter Hoff.

Examples

#Generate random data.
p <- c(2, 3, 4)
X <- array(stats::rnorm(prod(p)), dim = p)

#Calculate HOSVD.
hosvd_x <- hosvd(X)
S <- hosvd_x$S
U <- hosvd_x$U

#Recover X.
trim(X - atrans(S, U))
#> , , 1
#> 
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    0
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    0
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    0
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    0
#> 

#S is all-orthogonal.
trim(mat(S, 1) %*% t(mat(S, 1)))
#>          [,1]    [,2]
#> [1,] 16.97453 0.00000
#> [2,]  0.00000 3.74547
trim(mat(S, 2) %*% t(mat(S, 2)))
#>          [,1]     [,2]     [,3]
#> [1,] 10.70358 0.000000 0.000000
#> [2,]  0.00000 7.730355 0.000000
#> [3,]  0.00000 0.000000 2.286066
trim(mat(S, 3) %*% t(mat(S, 3)))
#>          [,1]     [,2]     [,3]     [,4]
#> [1,] 9.960677 0.000000 0.000000 0.000000
#> [2,] 0.000000 5.589668 0.000000 0.000000
#> [3,] 0.000000 0.000000 2.916853 0.000000
#> [4,] 0.000000 0.000000 0.000000 2.252804