The incredible higher-order polar decomposition (IHOP).

Mmm, pancakes.

ihop(X, itermax = 100, tol = 10^-9, print_diff = TRUE, mode_rep = NULL,
  use_sig = TRUE)

Arguments

X
An array of numerics.

itermax
An integer. The maximum number of iterations to perform during the optimization procedure.

tol
A numeric. The algorithm will stop when the Frobenius norm of the difference of core arrays between subsequent iterations is below tol (for use_sig = FALSE) or when the absolute difference between the ratio of subsequent values of sig and 1 is less than tol (for use_sig = TRUE).

print_diff
A logical. Should we print the updates of the algorithm?

mode_rep
A vector. Which component matrices should be set to be the identity?

use_sig
A logical. See tol.

Value

R A core array which, in combination with L, has certain orthogonality properties.

L A list of lower triangular matrices with unit Frobenius norm.

sig A numeric.

Details

This function will calculate the higher-order polar decomposition, a generalization of the polar decomposition to tensors. It generalizes a minimization formulation of the polar decomposition.

Given an array X, ihop will output L a list of lower triangular matrices with positive diagonal elements and unit Frobenius norm, R a core array with certain orthogonality properties, and sig a total variation parameter. We have that X is equal to sig * atrans(R, L) up to numerical precision.

t(solve(L[[i]])) %*% mat(R, i) will have orthonormal rows for all i.

For more details on the IHOP, see Gerard and Hoff (2016).

References

Gerard, D., & Hoff, P. (2016). A higher-order LQ decomposition for separable covariance models. Linear Algebra and its Applications, 505, 57-84. https://doi.org/10.1016/j.laa.2016.04.033 http://arxiv.org/pdf/1410.1094v1.pdf

Examples

#Generate random data.
p <- c(2, 3, 4)
X <- array(stats::rnorm(prod(p)), dim = p)

#Calculate IHOP.
ihop_x <- ihop(X)

#> Scale Diff =  0.2104518 
#>      Scale =  21.076 
#> 
#> Scale Diff =  0.01308105 
#>      Scale =  20.80386 
#> 
#> Scale Diff =  0.0005624172 
#>      Scale =  20.79217 
#> 
#> Scale Diff =  2.528993e-05 
#>      Scale =  20.79164 
#> 
#> Scale Diff =  1.165604e-06 
#>      Scale =  20.79162 
#> 
#> Scale Diff =  5.019131e-08 
#>      Scale =  20.79161 
#> 
#> Scale Diff =  1.915227e-09 
#>      Scale =  20.79161 
#> 
#> Scale Diff =  6.47169e-11 
#>      Scale =  20.79161 
#> 

R <- ihop_x$R
L <- ihop_x$L
sig <- ihop_x$sig

#Reconstruct X
trim(X - sig * atrans(R, L))

#> , , 1
#> 
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    0
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    0
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    0
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    0
#> 

#Orthogonality properties
ortho_1 <- t(solve(L[[1]])) %*% mat(R, 1)
trim(ortho_1 %*% t(ortho_1))

#>               [,1]          [,2]
#> [1,]  1.000002e+00 -1.027596e-06
#> [2,] -1.027596e-06  9.999988e-01

ortho_2 <- t(solve(L[[2]])) %*% mat(R, 2)
trim(ortho_2 %*% t(ortho_2))

#>               [,1]          [,2]          [,3]
#> [1,]  1.000001e+00  0.000000e+00 -3.113055e-06
#> [2,]  0.000000e+00  1.000000e+00 -2.520848e-06
#> [3,] -3.113055e-06 -2.520848e-06  9.999987e-01

ortho_3 <- t(solve(L[[3]])) %*% mat(R, 3)
trim(ortho_3 %*% t(ortho_3))

#>      [,1] [,2] [,3] [,4]
#> [1,]    1    0    0    0
#> [2,]    0    1    0    0
#> [3,]    0    0    1    0
#> [4,]    0    0    0    1