Calculate the likelihood ratio test statistic.

Calulate the likelihood ratio test statistic for Kronecker structured covariance models.

lrt_stat(sig_null, sig_alt, p)

Arguments

sig_null
A numeric. The MLE of the total variation parameter under the null (the standard deviation version).

sig_alt
A numeric. The MLE of the total variation parameter under the alternative (the standard deviation version).

p
A vector of integers. The dimension of the array.

Value

A numeric. The likelihood ratio test statistic.

Details

The LRT statistic is the exact same for all elliptically distributed Kronecker structured covariance models (not just the normal). The distribution of the likelihood ratio test statistic does change.

References

Gerard, D. C., & Hoff, P. D. (2014). A higher-order LQ decomposition for separable covariance models. arXiv preprint arXiv:1410.1094.

See also

holq for obtaining the MLE of the total variation parameter.

lrt_null_dist_dim_same for getting the null distribution of the likelihood ratio test statistic.