Calculate a truncated multiway Takemura estimator.

This function will 'average' Bayes rules given random rotations of the data array. This 'averaged' estimator has lower risk than the uniformly minimum risk equivariant estimator under a product group of lower triangular matrices. Truncated multiway Takemura's estimator is not equivariant with respect to this product group of lower triangular matrices, but it is an equivariant randomized estimator with respect to a product group of orthogonal matrices.

multiway_takemura(
  X,
  ortho_max = 2,
  mcmc_itermax = 1000,
  start_identity = FALSE,
  print_mcmc = FALSE,
  mode_rep = NULL
)

Arguments

X

An array. This is the data array.

ortho_max

An integer. The number of 'averagings' to perform.

mcmc_itermax

An integer. The number of iterations each MCMC should perform using equi_mcmc.

start_identity

Should each MCMC start their covariance matrices at the identity (TRUE) or at the sample covariance matrices (FALSE)?

print_mcmc

Should the output of the MCMC be printed to the screen (TRUE) or not (FALSE)?

mode_rep

A vector of integers. Which mode(s) are considered iid observations? Default is none.

Value

B A list of the truncated multiway Takemura's estimators for each component covariance matrix. Not their Cholesky square roots.

b Truncated multiway Takemura's estimator for the total variation parameter. The 'variance' form, not the 'standard devation' form.

Details

This function will (1) randomly rotate X along every mode, then (2) it will calculate the uniformly minimum risk equivariant estimator using equi_mcmc, then (3) it will 'average' these estimates.

References

Gerard, D., & Hoff, P. (2015). Equivariant minimax dominators of the MLE in the array normal model. Journal of Multivariate Analysis, 137, 32-49. doi:10.1016/j.jmva.2015.01.020

See also

Author

David Gerard.

Examples

# Simulate data.
p <- c(5, 5, 5)
X <- array(stats::rnorm(prod(p)), dim = p)
multi_out <- multiway_takemura(X, mode_rep = 3)
multi_out$b
#> [1] 0.9340344
trim(multi_out$B[[1]])
#>            [,1]        [,2]       [,3]        [,4]       [,5]
#> [1,]  1.6911401 -0.43010151 0.11701507  0.41027011 -0.2361179
#> [2,] -0.4301015  1.16308672 0.06532907 -0.01621828  0.0238104
#> [3,]  0.1170151  0.06532907 1.00984274  0.03490203  0.2395732
#> [4,]  0.4102701 -0.01621828 0.03490203  0.76661432 -0.2920952
#> [5,] -0.2361179  0.02381040 0.23957315 -0.29209522  1.0395229
trim(multi_out$B[[2]])
#>           [,1]        [,2]        [,3]       [,4]       [,5]
#> [1,] 1.2364194  0.18591914  0.48805777  0.7182665  0.1885822
#> [2,] 0.1859191  1.10393856 -0.03767331 -0.3347590 -0.3663760
#> [3,] 0.4880578 -0.03767331  1.21073763  0.1520050 -0.2113612
#> [4,] 0.7182665 -0.33475901  0.15200496  1.4037169  0.4635864
#> [5,] 0.1885822 -0.36637598 -0.21136120  0.4635864  1.2687095
trim(multi_out$B[[3]])
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    1    0    0    0    0
#> [2,]    0    1    0    0    0
#> [3,]    0    0    1    0    0
#> [4,]    0    0    0    1    0
#> [5,]    0    0    0    0    1