Array indices.

Generates indices corresponding to subarrays.

arrIndices(saidx)

Arguments

saidx

either a vector of the dimensions of a potential array, or a list of the indices in the subarray.

Details

This function generates a matrix corresponding to all combinations of a list of indices, to be used in subsetting arrays.

Author

Peter Hoff.

Examples

# all indices of an array
arrIndices(c(4, 3, 2))
#>       dim1 dim2 dim3
#>  [1,]    1    1    1
#>  [2,]    2    1    1
#>  [3,]    3    1    1
#>  [4,]    4    1    1
#>  [5,]    1    2    1
#>  [6,]    2    2    1
#>  [7,]    3    2    1
#>  [8,]    4    2    1
#>  [9,]    1    3    1
#> [10,]    2    3    1
#> [11,]    3    3    1
#> [12,]    4    3    1
#> [13,]    1    1    2
#> [14,]    2    1    2
#> [15,]    3    1    2
#> [16,]    4    1    2
#> [17,]    1    2    2
#> [18,]    2    2    2
#> [19,]    3    2    2
#> [20,]    4    2    2
#> [21,]    1    3    2
#> [22,]    2    3    2
#> [23,]    3    3    2
#> [24,]    4    3    2
# indices of a subarray
arrIndices(list(c(1, 3), c(4, 5), c(2, 3, 6)))
#>       dim1 dim2 dim3
#>  [1,]    1    4    2
#>  [2,]    3    4    2
#>  [3,]    1    5    2
#>  [4,]    3    5    2
#>  [5,]    1    4    3
#>  [6,]    3    4    3
#>  [7,]    1    5    3
#>  [8,]    3    5    3
#>  [9,]    1    4    6
#> [10,]    3    4    6
#> [11,]    1    5    6
#> [12,]    3    5    6