Calculate the incredible SVD (ISVD).

The ISVD is a generalization of the SVD to tensors. It is derived from the incredible HOLQ.

get_isvd(x_holq)

Arguments

x_holq
The output from holq.

Value

l A numeric.

U A list of orthogonal matrices.

D A list of diagonal matrices with positive diagonal entries and unit determinant. The diagonal entries are in descending order.

V A scaled all-orthonormal array.

Details

Let sig * atrans(Z, L) be the HOLQ of X. Then the ISVD calculates the SVD of each L[[i]], call it U[[i]] %*% D[[i]] %*% t(W[[i]]). It then returns l = sig, U, D, and V = atrans(Z, W). These values have the property that X is equal to l * atrans(atrans(V, D), U), up to numerical precision. V is also scaled all-orthonormal.

For more details on the ISVD, see Gerard and Hoff (2014).

References

Gerard, D. C., & Hoff, P. D. (2014). A higher-order LQ decomposition for separable covariance models. arXiv preprint arXiv:1410.1094.

Examples

#Generate random data.
p <- c(4,4,4)
X <- array(stats::rnorm(prod(p)), dim = p)

#Calculate HOLQ, then ISVD
holq_x <- holq(X)

#> Scale Diff =  0.197 
#>      Scale =  6.66 
#> 
#> Scale Diff =  0.04906 
#>      Scale =  6.349 
#> 
#> Scale Diff =  0.01107 
#>      Scale =  6.279 
#> 
#> Scale Diff =  0.004511 
#>      Scale =  6.251 
#> 
#> Scale Diff =  0.001877 
#>      Scale =  6.239 
#> 
#> Scale Diff =  0.0007693 
#>      Scale =  6.235 
#> 
#> Scale Diff =  0.0003119 
#>      Scale =  6.233 
#> 
#> Scale Diff =  0.0001263 
#>      Scale =  6.232 
#> 
#> Scale Diff =  5.15e-05 
#>      Scale =  6.232 
#> 
#> Scale Diff =  2.116e-05 
#>      Scale =  6.231 
#> 
#> Scale Diff =  8.748e-06 
#>      Scale =  6.231 
#> 
#> Scale Diff =  3.628e-06 
#>      Scale =  6.231 
#> 
#> Scale Diff =  1.506e-06 
#>      Scale =  6.231 
#> 
#> Scale Diff =  6.254e-07 
#>      Scale =  6.231 
#> 
#> Scale Diff =  2.596e-07 
#>      Scale =  6.231 
#> 
#> Scale Diff =  1.079e-07 
#>      Scale =  6.231 
#> 
#> Scale Diff =  4.489e-08 
#>      Scale =  6.231 
#> 
#> Scale Diff =  1.874e-08 
#>      Scale =  6.231 
#> 
#> Scale Diff =  7.852e-09 
#>      Scale =  6.231 
#> 
#> Scale Diff =  3.308e-09 
#>      Scale =  6.231 
#> 
#> Scale Diff =  1.402e-09 
#>      Scale =  6.231 
#> 
#> Scale Diff =  5.977e-10 
#>      Scale =  6.231 
#> 

isvd_x <- get_isvd(holq_x)
l <- isvd_x$l
U <- isvd_x$U
D <- isvd_x$D
V <- isvd_x$V

#Recover X
trim(X - l * atrans(atrans(V, D), U))

#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    0    0    0    0
#> [2,]    0    0    0    0
#> [3,]    0    0    0    0
#> [4,]    0    0    0    0
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    0    0    0    0
#> [2,]    0    0    0    0
#> [3,]    0    0    0    0
#> [4,]    0    0    0    0
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    0    0    0    0
#> [2,]    0    0    0    0
#> [3,]    0    0    0    0
#> [4,]    0    0    0    0
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]    0    0    0    0
#> [2,]    0    0    0    0
#> [3,]    0    0    0    0
#> [4,]    0    0    0    0
#> 

#V is scaled all-orthonormal
trim(mat(V, 1) %*% t(mat(V, 1)), epsilon = 10^-5)

#>      [,1] [,2] [,3] [,4]
#> [1,] 0.25 0.00 0.00 0.00
#> [2,] 0.00 0.25 0.00 0.00
#> [3,] 0.00 0.00 0.25 0.00
#> [4,] 0.00 0.00 0.00 0.25

trim(mat(V, 2) %*% t(mat(V, 2)), epsilon = 10^-5)

#>      [,1] [,2] [,3] [,4]
#> [1,] 0.25 0.00 0.00 0.00
#> [2,] 0.00 0.25 0.00 0.00
#> [3,] 0.00 0.00 0.25 0.00
#> [4,] 0.00 0.00 0.00 0.25

trim(mat(V, 3) %*% t(mat(V, 3)), epsilon = 10^-5)

#>      [,1] [,2] [,3] [,4]
#> [1,] 0.25 0.00 0.00 0.00
#> [2,] 0.00 0.25 0.00 0.00
#> [3,] 0.00 0.00 0.25 0.00
#> [4,] 0.00 0.00 0.00 0.25