Calculate a truncated multiway Takemura estimator.

This function will 'average' Bayes rules given random rotations of the data array. This 'averaged' estimator has lower risk than the uniformly minimum risk equivariant estimator under a product group of lower triangular matrices. Truncated multiway Takemura's estimator is not equivariant with respect to this product group of lower triangular matrices, but it is an equivariant randomized estimator with respect to a product group of orthogonal matrices.

multiway_takemura(X, ortho_max = 2, mcmc_itermax = 1000,
  start_identity = FALSE, print_mcmc = FALSE, mode_rep = NULL)

Arguments

X
An array. This is the data array.

ortho_max
An integer. The number of 'averagings' to perform.

mcmc_itermax
An integer. The number of iterations each MCMC should perform using equi_mcmc.

start_identity
Should each MCMC start their covariance matrices at the identity (TRUE) or at the sample covariance matrices (FALSE)?

print_mcmc
Should the output of the MCMC be printed to the screen (TRUE) or not (FALSE)?

mode_rep
A vector of integers. Which mode(s) are considered iid observations? Default is none.

Value

B A list of the truncated multiway Takemura's estimators for each component covariance matrix. Not their Cholesky square roots.

b Truncated multiway Takemura's estimator for the total variation parameter. The 'variance' form, not the 'standard devation' form.

Details

This function will (1) randomly rotate X along every mode, then (2) it will calculate the uniformly minimum risk equivariant estimator using equi_mcmc, then (3) it will 'average' these estimates.

References

Gerard, D., & Hoff, P. (2015). Equivariant minimax dominators of the MLE in the array normal model. Journal of Multivariate Analysis, 137, 32-49.

See also

equi_mcmc, random_ortho.

Examples

# Simulate data.
p <- c(5, 5, 5)
X <- array(stats::rnorm(prod(p)), dim = p)
multi_out <- multiway_takemura(X, mode_rep = 3)
multi_out$b

#> [1] 0.8036

trim(multi_out$B[[1]])

#>           [,1]     [,2]    [,3]    [,4]      [,5]
#> [1,]  0.391260 -0.19722  0.2184  0.0735 -0.001292
#> [2,] -0.197217  1.31230 -0.1964 -0.3053 -0.089305
#> [3,]  0.218418 -0.19641  1.6119 -0.2495 -0.342085
#> [4,]  0.073498 -0.30530 -0.2495  1.8078  0.671265
#> [5,] -0.001292 -0.08931 -0.3421  0.6713  1.139266

trim(multi_out$B[[2]])

#>          [,1]     [,2]     [,3]     [,4]     [,5]
#> [1,]  0.83558  0.27839 -0.13689  0.03655 -0.37103
#> [2,]  0.27839  1.00517  0.07682 -0.15349 -0.26208
#> [3,] -0.13689  0.07682  1.22909  0.33125  0.41176
#> [4,]  0.03655 -0.15349  0.33125  1.31450 -0.05019
#> [5,] -0.37103 -0.26208  0.41176 -0.05019  1.26379

trim(multi_out$B[[3]])

#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    1    0    0    0    0
#> [2,]    0    1    0    0    0
#> [3,]    0    0    1    0    0
#> [4,]    0    0    0    1    0
#> [5,]    0    0    0    0    1