Update for total variation parameter in equi_mcmc.

Samples from the square root of an inverse-gamma.

sample_sig(X, phi_inv)

Arguments

X

An array. The tensor data.

phi_inv

A list of the current values of inverse of the lower-triangular Cholesky square root of the the component covariance matrices. This is equivalent to the transpose of the upper-triangular Cholesky square root of the inverse component covariance matrices.

phi_inv[[i]] is a lower triangluar matrix where solve(phi_inv[[i]]) %*% t(solve(phi_inv[[i]])) is the current estimate of the \(i\)th component covariance matrix.

Value

A numeric. The update for the total variation parameter in the MCMC implemented in equi_bayes.

Details

This function provides a Gibbs update for the total variation parameter from the MCMC implemented in equi_mcmc. This corresponds to the square root of an inverse-gamma distributed random variable whose parameters depend on the data and the component covariance matrices. Roughly, this is the update for the standard deviation, not the variance.

References

Gerard, D., & Hoff, P. (2015). Equivariant minimax dominators of the MLE in the array normal model. Journal of Multivariate Analysis, 137, 32-49. doi:10.1016/j.jmva.2015.01.020

See also

Author

David Gerard.